requirements.txt
Flask==3.0.0
Werkzeug==3.0.1

# PDF to Word conversion with table support
pdf2docx==0.5.8

# Word document manipulation
python-docx==1.1.0

# Fallback table extraction from PDFs
pdfplumber==0.11.0

# Additional dependencies
Pillow==10.1.0
PyMuPDF==1.23.8

# PDF generation from DOCX
docx2pdf==0.1.8

# AI Summary
ollama==0.1.6

# Batch processing
docxcompose==1.4.0

# Database Connection (Added for API Tracking)
mysql-connector-python==9.6.0